Test that beans from a configuration bean get passed to bean methods
This tests that the beans created by a configuration bean can be used to
autowire the bean creation methods of beans that are created by the base
configuration class.

That is a pretty long winded way of saying that the use of configuration
beans should not lead to any issues with unresolvable beans. Provided
that the configuration beans can be created their inner beans should be
available for creation and use by other unrelated bean methods.